requirements.txt
streamlit==1.38.0
requests==2.32.3
pandas==2.2.2
mistralai==1.0.1
plotly==5.24.0
python-dotenv==1.0.1
numpy==1.26.4
